+++ src/imio/smartweb/core/tests/test_rest.py
@@ -18,7 +18,6 @@ from plone.app.testing import TEST_USER_ID
 from plone.app.testing import TEST_USER_PASSWORD
 from plone.namedfile.file import NamedBlobImage
 from plone.restapi.testing import RelativeSession
-from unittest import mock
 from unittest.mock import patch
 from urllib.parse import urlparse
 from urllib.parse import parse_qs
@@ -29,7 +28,6 @@ from ZPublisher.pubevents import PubStart
 from imio.smartweb.core.viewlets.ogptags import OgpTagsViewlet
 
 import json
-import os
 import requests_mock
 import transaction
 
@@ -444,8 +442,8 @@ class SectionsFunctionalTest(ImioSmartwebTestCase):
             "metadata_fields=UID&"
             "sort_on=effective&"
             "sort_order=descending&"
-            "b_size=20&"
             "fullobjects=1&"
+            "b_size=20&"
             "translated_in_en=1".format(self.rest_news.selected_news_folder),
         )
         m.get(url, text=json.dumps({}))