+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.3.1 (unreleased)
 ------------------
 
+- WEB-4165: Refactoring the generation of sitemap.xml.gz for search engines and modifying the site map for user display
+  [boulch]
+
 - WEB-4165 : Fixed sitemap.xml.gz generation with auth sources entries
   [boulch]